[WIP] Feedback on the PR
