+++ src/python/grpcio_tests/tests_aio/tests.json
@@ -9,7 +9,6 @@
   "unit.call_test.TestStreamUnaryCall",
   "unit.call_test.TestUnaryStreamCall",
   "unit.call_test.TestUnaryUnaryCall",
-  "unit.call_test.TestUnaryUnarySecureCall",
   "unit.channel_argument_test.TestChannelArgument",
   "unit.channel_ready_test.TestChannelReady",
   "unit.channel_test.TestChannel",
@@ -19,10 +18,11 @@
   "unit.compression_test.TestCompression",
   "unit.connectivity_test.TestConnectivityState",
   "unit.done_callback_test.TestDoneCallback",
-  "unit.init_test.TestInsecureChannel",
-  "unit.init_test.TestSecureChannel",
+  "unit.init_test.TestChannel",
+  "unit.interceptor_test.TestInterceptedUnaryUnaryCall",
+  "unit.interceptor_test.TestUnaryUnaryClientInterceptor",
   "unit.metadata_test.TestMetadata",
-  "unit.server_interceptor_test.TestServerInterceptor",
+  "unit.secure_call_test.TestUnaryUnarySecureCall",
   "unit.server_test.TestServer",
   "unit.timeout_test.TestTimeout",
   "unit.wait_for_ready_test.TestWaitForReady"
