+++ src/python/grpcio_tests/tests_aio/unit/call_test.py
@@ -34,7 +34,6 @@ _LOCAL_CANCEL_DETAILS_EXPECTATION = 'Locally cancelled by application!'
 _RESPONSE_INTERVAL_US = test_constants.SHORT_TIMEOUT * 1000 * 1000
 _UNREACHABLE_TARGET = '0.1:1111'
 _INFINITE_INTERVAL_US = 2**31 - 1
-_SERVER_HOST_OVERRIDE = 'foo.test.google.fr'
 
 
 class _MulticallableTestMixin():
@@ -42,33 +41,6 @@ class _MulticallableTestMixin():
     async def setUp(self):
         address, self._server = await start_test_server()
         self._channel = aio.insecure_channel(address)
-        self._stub = test_pb2_grpc.TestServiceStub(self._channel)
-
-    async def tearDown(self):
-        await self._channel.close()
-        await self._server.stop(None)
-
-
-
-class _SecureCallMixin:
-    """A Mixin to run the call tests over a secure channel."""
-
-    async def setUp(self):
-        server_credentials = grpc.ssl_server_credentials([
-            (resources.private_key(), resources.certificate_chain())
-        ])
-        channel_credentials = grpc.ssl_channel_credentials(
-            resources.test_root_certificates())
-
-        self._server_address, self._server = await start_test_server(
-            secure=True, server_credentials=server_credentials)
-        channel_options = (
-            (
-                'grpc.ssl_target_name_override',
-                _SERVER_HOST_OVERRIDE,
-            ),
-        )
-        self._channel = aio.secure_channel(self._server_address, channel_credentials, channel_options)
         self._stub = test_pb2_grpc.TestServiceStub(self._channel)
 
     async def tearDown(self):
@@ -236,17 +208,11 @@ class TestUnaryUnaryCall(_MulticallableTestMixin, AioTestBase):
             grpc.access_token_call_credentials("abc"),
             grpc.access_token_call_credentials("def"),
         )
-        with self.assertRaisesRegex(RuntimeError, "Call credentials are only valid on secure channels"):
-            self._stub.UnaryCall(messages_pb2.SimpleRequest(), credentials=call_credentials)
-
-
-class TestUnaryUnarySecureCall(_SecureCallMixin, AioTestBase):
-    """Calls made over a secure channel."""
-    async def test_call_ok_with_credentials(self):
-        call = self._stub.UnaryCall(messages_pb2.SimpleRequest())
-        response = await call
-        self.assertIsInstance(response, messages_pb2.SimpleResponse)
-        self.assertEqual(await call.code(), grpc.StatusCode.OK)
+        with self.assertRaisesRegex(
+                RuntimeError,
+                "Call credentials are only valid on secure channels"):
+            self._stub.UnaryCall(messages_pb2.SimpleRequest(),
+                                 credentials=call_credentials)
 
 
 class TestUnaryStreamCall(_MulticallableTestMixin, AioTestBase):
@@ -601,6 +567,7 @@ _STREAM_OUTPUT_REQUEST_ONE_RESPONSE.response_parameters.append(
 
 
 class TestStreamStreamCall(_MulticallableTestMixin, AioTestBase):
+
     async def test_cancel(self):
         # Invokes the actual RPC
         call = self._stub.FullDuplexCall()
